feat!: represent money as integer cents
Every monetary value was a float, and this package adds money up in
loops: getSubTotal() accumulates one line total per item, and totals,
VAT at 6/12/21% and percentage discounts are all computed by
accumulation. Float drifts exactly there -- 0.1 + 0.2 + 0.3 is not 0.6,
while 10 + 20 + 30 is always 60.

Money is now an int counting cents across the contracts, the domain, the
models and the schema. The only floats left in src/ are stock quantities,
which are not money and belong to sc-11173.

Integers do not remove rounding, they concentrate it, so the rule is
stated once in Tnt\Ecommerce\Money and documented in the README:

1. Round half away from zero. 21% of 50 is 10.5, and becomes 11. Banker's
   rounding reads as arbitrary on a single invoice line.
2. Round once, on the smallest amount the rate genuinely applies to.
   Per-line VAT rounds per line; a cart-level percentage discount rounds
   once on the subtotal. Totals are integer sums of already-rounded
   amounts, never a rate applied to a total.

The second half is a real choice: two lines of 1250 at 21% give 263 + 263
= 526 per line but 525 on the total. The package takes 526, because each
line is separately printed and separately checkable, so the total must be
the sum of the printed lines. Asserted explicitly in the tests.

Money::percentageOf reduces the rate to its smallest whole fraction
before multiplying and divides with integer arithmetic, so no float
enters the money path at all and results stay exact past 2^53 cents.

Money columns become bigint(20) rather than int(11). int stops at
2,147,483,647 cents, which is narrower than the decimal(10,2) it
replaces, so it would have silently truncated large orders instead of
refusing them. Verified against the MySQL 8.0.34 container that bigint
round-trips the full signed 64-bit range.

Tests 77 -> 93. PHPStan baseline unchanged at 33 entries, none in any
file touched here. The 72 unit tests still pass with the database
container stopped.

BREAKING CHANGE: every money-carrying contract now returns int cents
instead of float, and the order money columns are bigint.

[sc-11170]

+++ src/Cart/Cart.php
@@ -114,9 +114,9 @@ class Cart implements CartInterface, TotalingInterface
     }
 
     /**
-     * @return float
+     * @return int
      */
-    public function getFulfillmentCost(): float
+    public function getFulfillmentCost(): int
     {
         $fulfill = $this->getFulfillment();
 
@@ -179,9 +179,14 @@ class Cart implements CartInterface, TotalingInterface
     }
 
     /**
-     * @return float
+     * The sum of the line totals, in cents.
+     *
+     * Every term is an integer, so the accumulation below is exact however many
+     * lines it runs over — which is the reason money is not a float here.
+     *
+     * @return int
      */
-    public function getSubTotal(): float
+    public function getSubTotal(): int
     {
         $cost = 0;
 
@@ -193,9 +198,9 @@ class Cart implements CartInterface, TotalingInterface
     }
 
     /**
-     * @return float
+     * @return int
      */
-    public function getTotal(): float
+    public function getTotal(): int
     {
         return $this->getSubTotal() +
             $this->getFulfillmentCost() -
@@ -203,9 +208,9 @@ class Cart implements CartInterface, TotalingInterface
     }
 
     /**
-     * @return float
+     * @return int
      */
-    public function getReduction(): float
+    public function getReduction(): int
     {
         $coupon = $this->getCoupon();
